+++ src/contracts/abi/chain.abi.ts
@@ -4,21 +4,21 @@ export const chainAbi = [
       {
         internalType: 'address',
         name: '_contractRegistry',
-        type: 'address',
+        type: 'address'
       },
       {
         internalType: 'uint16',
         name: '_padding',
-        type: 'uint16',
+        type: 'uint16'
       },
       {
         internalType: 'uint16',
         name: '_requiredSignatures',
-        type: 'uint16',
-      },
+        type: 'uint16'
+      }
     ],
     stateMutability: 'nonpayable',
-    type: 'constructor',
+    type: 'constructor'
   },
   {
     anonymous: false,
@@ -27,29 +27,29 @@ export const chainAbi = [
         indexed: true,
         internalType: 'address',
         name: 'minter',
-        type: 'address',
+        type: 'address'
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'blockId',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'staked',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'power',
-        type: 'uint256',
-      },
+        type: 'uint256'
+      }
     ],
     name: 'LogMint',
-    type: 'event',
+    type: 'event'
   },
   {
     anonymous: false,
@@ -58,17 +58,17 @@ export const chainAbi = [
         indexed: true,
         internalType: 'address',
         name: 'executor',
-        type: 'address',
+        type: 'address'
       },
       {
         indexed: false,
         internalType: 'uint16',
         name: 'timePadding',
-        type: 'uint16',
-      },
+        type: 'uint16'
+      }
     ],
     name: 'LogPadding',
-    type: 'event',
+    type: 'event'
   },
   {
     anonymous: false,
@@ -77,23 +77,23 @@ export const chainAbi = [
         indexed: true,
         internalType: 'uint256',
         name: 'blockId',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         indexed: true,
         internalType: 'address',
         name: 'voter',
-        type: 'address',
+        type: 'address'
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'vote',
-        type: 'uint256',
-      },
+        type: 'uint256'
+      }
     ],
     name: 'LogVoter',
-    type: 'event',
+    type: 'event'
   },
   {
     anonymous: false,
@@ -102,17 +102,17 @@ export const chainAbi = [
         indexed: true,
         internalType: 'address',
         name: 'previousOwner',
-        type: 'address',
+        type: 'address'
       },
       {
         indexed: true,
         internalType: 'address',
         name: 'newOwner',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     name: 'OwnershipTransferred',
-    type: 'event',
+    type: 'event'
   },
   {
     inputs: [],
@@ -121,35 +121,42 @@ export const chainAbi = [
       {
         internalType: 'bytes',
         name: '',
-        type: 'bytes',
-      },
+        type: 'bytes'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
+        name: '_blockId',
+        type: 'uint256'
+      }
     ],
     name: 'blocks',
     outputs: [
       {
-        internalType: 'bytes32',
-        name: 'root',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'uint32',
-        name: 'dataTimestamp',
-        type: 'uint32',
-      },
+        components: [
+          {
+            internalType: 'bytes32',
+            name: 'root',
+            type: 'bytes32'
+          },
+          {
+            internalType: 'uint32',
+            name: 'dataTimestamp',
+            type: 'uint32'
+          }
+        ],
+        internalType: 'struct BaseChain.Block',
+        name: '',
+        type: 'tuple'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -158,11 +165,11 @@ export const chainAbi = [
       {
         internalType: 'uint32',
         name: '',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -171,40 +178,40 @@ export const chainAbi = [
       {
         internalType: 'uint32',
         name: '',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'bytes',
         name: '_data',
-        type: 'bytes',
+        type: 'bytes'
       },
       {
         internalType: 'uint256',
         name: '_offset',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         internalType: 'uint256',
         name: '_items',
-        type: 'uint256',
-      },
+        type: 'uint256'
+      }
     ],
     name: 'bytesToBytes32Array',
     outputs: [
       {
         internalType: 'bytes32[]',
         name: '',
-        type: 'bytes32[]',
-      },
+        type: 'bytes32[]'
+      }
     ],
     stateMutability: 'pure',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -213,35 +220,35 @@ export const chainAbi = [
       {
         internalType: 'contract IRegistry',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '',
-        type: 'bytes32',
-      },
+        type: 'bytes32'
+      }
     ],
     name: 'fcds',
     outputs: [
       {
         internalType: 'uint224',
         name: 'value',
-        type: 'uint224',
+        type: 'uint224'
       },
       {
         internalType: 'uint32',
         name: 'dataTimestamp',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -250,116 +257,116 @@ export const chainAbi = [
       {
         internalType: 'uint32',
         name: '',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_timestamp',
-        type: 'uint256',
-      },
+        type: 'uint256'
+      }
     ],
     name: 'getBlockIdAtTimestamp',
     outputs: [
       {
         internalType: 'uint32',
         name: '',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint32',
         name: '_blockId',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     name: 'getBlockRoot',
     outputs: [
       {
         internalType: 'bytes32',
         name: '',
-        type: 'bytes32',
-      },
+        type: 'bytes32'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint32',
         name: '_blockId',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     name: 'getBlockTimestamp',
     outputs: [
       {
         internalType: 'uint32',
         name: '',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '_key',
-        type: 'bytes32',
-      },
+        type: 'bytes32'
+      }
     ],
     name: 'getCurrentValue',
     outputs: [
       {
         internalType: 'uint256',
         name: 'value',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         internalType: 'uint256',
         name: 'timestamp',
-        type: 'uint256',
-      },
+        type: 'uint256'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'bytes32[]',
         name: '_keys',
-        type: 'bytes32[]',
-      },
+        type: 'bytes32[]'
+      }
     ],
     name: 'getCurrentValues',
     outputs: [
       {
         internalType: 'uint256[]',
         name: 'values',
-        type: 'uint256[]',
+        type: 'uint256[]'
       },
       {
         internalType: 'uint32[]',
         name: 'timestamps',
-        type: 'uint32[]',
-      },
+        type: 'uint32[]'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -368,11 +375,11 @@ export const chainAbi = [
       {
         internalType: 'uint32',
         name: '',
-        type: 'uint32',
-      },
+        type: 'uint32'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -381,54 +388,54 @@ export const chainAbi = [
       {
         internalType: 'address',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_timestamp',
-        type: 'uint256',
-      },
+        type: 'uint256'
+      }
     ],
     name: 'getLeaderAddressAtTime',
     outputs: [
       {
         internalType: 'address',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_numberOfValidators',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         internalType: 'uint256',
         name: '_timestamp',
-        type: 'uint256',
-      },
+        type: 'uint256'
+      }
     ],
     name: 'getLeaderIndex',
     outputs: [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256',
-      },
+        type: 'uint256'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -437,11 +444,11 @@ export const chainAbi = [
       {
         internalType: 'bytes32',
         name: '',
-        type: 'bytes32',
-      },
+        type: 'bytes32'
+      }
     ],
     stateMutability: 'pure',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -450,11 +457,11 @@ export const chainAbi = [
       {
         internalType: 'address',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -463,85 +470,98 @@ export const chainAbi = [
       {
         internalType: 'uint256',
         name: 'blockNumber',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         internalType: 'uint16',
         name: 'timePadding',
-        type: 'uint16',
+        type: 'uint16'
       },
       {
         internalType: 'uint32',
         name: 'lastDataTimestamp',
-        type: 'uint32',
+        type: 'uint32'
       },
       {
         internalType: 'uint32',
         name: 'lastBlockId',
-        type: 'uint32',
+        type: 'uint32'
       },
       {
         internalType: 'address',
         name: 'nextLeader',
-        type: 'address',
+        type: 'address'
       },
       {
         internalType: 'uint32',
         name: 'nextBlockId',
-        type: 'uint32',
+        type: 'uint32'
       },
       {
         internalType: 'address[]',
         name: 'validators',
-        type: 'address[]',
+        type: 'address[]'
       },
       {
         internalType: 'uint256[]',
         name: 'powers',
-        type: 'uint256[]',
+        type: 'uint256[]'
       },
       {
         internalType: 'string[]',
         name: 'locations',
-        type: 'string[]',
+        type: 'string[]'
       },
       {
         internalType: 'uint256',
         name: 'staked',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         internalType: 'uint16',
         name: 'minSignatures',
-        type: 'uint16',
-      },
+        type: 'uint16'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'bytes',
         name: '_key',
-        type: 'bytes',
+        type: 'bytes'
       },
       {
         internalType: 'bytes',
         name: '_value',
-        type: 'bytes',
-      },
+        type: 'bytes'
+      }
     ],
     name: 'hashLeaf',
     outputs: [
       {
         internalType: 'bytes32',
         name: '',
-        type: 'bytes32',
-      },
+        type: 'bytes32'
+      }
     ],
     stateMutability: 'pure',
-    type: 'function',
+    type: 'function'
+  },
+  {
+    inputs: [],
+    name: 'isForeign',
+    outputs: [
+      {
+        internalType: 'bool',
+        name: '',
+        type: 'bool'
+      }
+    ],
+    stateMutability: 'pure',
+    type: 'function'
   },
   {
     inputs: [],
@@ -550,11 +570,11 @@ export const chainAbi = [
       {
         internalType: 'address',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -563,52 +583,59 @@ export const chainAbi = [
       {
         internalType: 'uint16',
         name: '',
-        type: 'uint16',
-      },
+        type: 'uint16'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '_affidavit',
-        type: 'bytes32',
+        type: 'bytes32'
       },
       {
         internalType: 'uint8',
         name: '_v',
-        type: 'uint8',
+        type: 'uint8'
       },
       {
         internalType: 'bytes32',
         name: '_r',
-        type: 'bytes32',
+        type: 'bytes32'
       },
       {
         internalType: 'bytes32',
         name: '_s',
-        type: 'bytes32',
-      },
+        type: 'bytes32'
+      }
     ],
     name: 'recoverSigner',
     outputs: [
       {
         internalType: 'address',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'pure',
-    type: 'function',
+    type: 'function'
+  },
+  {
+    inputs: [],
+    name: 'register',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function'
   },
   {
     inputs: [],
     name: 'renounceOwnership',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -617,24 +644,43 @@ export const chainAbi = [
       {
         internalType: 'uint16',
         name: '',
-        type: 'uint16',
-      },
+        type: 'uint16'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint16',
         name: '_padding',
-        type: 'uint16',
-      },
+        type: 'uint16'
+      }
     ],
     name: 'setPadding',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function',
+    type: 'function'
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256'
+      }
+    ],
+    name: 'squashedRoots',
+    outputs: [
+      {
+        internalType: 'bytes32',
+        name: '',
+        type: 'bytes32'
+      }
+    ],
+    stateMutability: 'view',
+    type: 'function'
   },
   {
     inputs: [],
@@ -643,11 +689,11 @@ export const chainAbi = [
       {
         internalType: 'contract IStakingBank',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -656,54 +702,54 @@ export const chainAbi = [
       {
         internalType: 'contract IStakingBank',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint32',
         name: '_dataTimestamp',
-        type: 'uint32',
+        type: 'uint32'
       },
       {
         internalType: 'bytes32',
         name: '_root',
-        type: 'bytes32',
+        type: 'bytes32'
       },
       {
         internalType: 'bytes32[]',
         name: '_keys',
-        type: 'bytes32[]',
+        type: 'bytes32[]'
       },
       {
         internalType: 'uint256[]',
         name: '_values',
-        type: 'uint256[]',
+        type: 'uint256[]'
       },
       {
         internalType: 'uint8[]',
         name: '_v',
-        type: 'uint8[]',
+        type: 'uint8[]'
       },
       {
         internalType: 'bytes32[]',
         name: '_r',
-        type: 'bytes32[]',
+        type: 'bytes32[]'
       },
       {
         internalType: 'bytes32[]',
         name: '_s',
-        type: 'bytes32[]',
-      },
+        type: 'bytes32[]'
+      }
     ],
     name: 'submit',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [],
@@ -712,120 +758,127 @@ export const chainAbi = [
       {
         internalType: 'contract ERC20',
         name: '',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: 'newOwner',
-        type: 'address',
-      },
+        type: 'address'
+      }
     ],
     name: 'transferOwnership',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function',
+    type: 'function'
+  },
+  {
+    inputs: [],
+    name: 'unregister',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'bytes32[]',
         name: '_proof',
-        type: 'bytes32[]',
+        type: 'bytes32[]'
       },
       {
         internalType: 'bytes32',
         name: '_root',
-        type: 'bytes32',
+        type: 'bytes32'
       },
       {
         internalType: 'bytes32',
         name: '_leaf',
-        type: 'bytes32',
-      },
+        type: 'bytes32'
+      }
     ],
     name: 'verifyProof',
     outputs: [
       {
         internalType: 'bool',
         name: '',
-        type: 'bool',
-      },
+        type: 'bool'
+      }
     ],
     stateMutability: 'pure',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_blockId',
-        type: 'uint256',
+        type: 'uint256'
       },
       {
         internalType: 'bytes32[]',
         name: '_proof',
-        type: 'bytes32[]',
+        type: 'bytes32[]'
       },
       {
         internalType: 'bytes',
         name: '_key',
-        type: 'bytes',
+        type: 'bytes'
       },
       {
         internalType: 'bytes',
         name: '_value',
-        type: 'bytes',
-      },
+        type: 'bytes'
+      }
     ],
     name: 'verifyProofForBlock',
     outputs: [
       {
         internalType: 'bool',
         name: '',
-        type: 'bool',
-      },
+        type: 'bool'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
+    type: 'function'
   },
   {
     inputs: [
       {
         internalType: 'uint32[]',
         name: '_blockIds',
-        type: 'uint32[]',
+        type: 'uint32[]'
       },
       {
         internalType: 'bytes',
         name: '_proofs',
-        type: 'bytes',
+        type: 'bytes'
       },
       {
         internalType: 'uint256[]',
         name: '_proofItemsCounter',
-        type: 'uint256[]',
+        type: 'uint256[]'
       },
       {
         internalType: 'bytes32[]',
         name: '_leaves',
-        type: 'bytes32[]',
-      },
+        type: 'bytes32[]'
+      }
     ],
     name: 'verifyProofs',
     outputs: [
       {
         internalType: 'bool[]',
         name: 'results',
-        type: 'bool[]',
-      },
+        type: 'bool[]'
+      }
     ],
     stateMutability: 'view',
-    type: 'function',
-  },
+    type: 'function'
+  }
 ];