+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Changed
+- adjust `SortedMerkleTree` to support squashed root 
 
 ## [5.0.0] - 2021-09-29
 ### Added